requirements.txt
crewai
python-dotenv
requests
pyTelegramBotAPI
sqlalchemy
cloudinary
